+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Note the the latest version is usually work in progress and may have not yet bee
 ## Added
 
 - `--confirm` flag to deletes / updates
+- `envelope env print-script --type unexport`
 
 ## Changed